- Add call to abs() as the GEMM output may differ slightly due to floating point accuracy and result in a negative distance which returns NaN if sqrt() is applied to it.   - Increase tolerance for CDist test due to the expected value being calculated without using matmul.   - Update math::Gemm to use the type for alpha and beta instead of hardcoding to float. Matches the GemmEx definition.

+++ onnxruntime/core/util/math_cpu.cc
@@ -86,7 +86,7 @@ void Gemm<float, ThreadPool>(const CBLAS_TRANSPOSE TransA, const CBLAS_TRANSPOSE
 
 template <>
 void Gemm<double, ThreadPool>(const CBLAS_TRANSPOSE TransA, const CBLAS_TRANSPOSE TransB, const int64_t M,
-                              const int64_t N, const int64_t K, float alpha, const double* A, const double* B, float beta,
+                              const int64_t N, const int64_t K, double alpha, const double* A, const double* B, double beta,
                               double* C, ThreadPool* threadpool) {
   int lda = static_cast<int>((TransA == CblasNoTrans) ? K : M);
   int ldb = static_cast<int>((TransB == CblasNoTrans) ? N : K);
